* Added symbolic link to serial device * The symbols are resampled (x5) and filtered using square-root-raised-cosine filter

+++ IO.h
@@ -38,7 +38,7 @@ public:
 
 	void process();
 
-	void write(MMDVM_STATE mode, int16_t* samples, uint16_t length, const uint8_t* control = NULL);
+	void write(MMDVM_STATE mode, float* samples, uint16_t length, const uint8_t* control = NULL);
 
 	uint16_t getSpace() const;
